core: move grandpa::is_descendent_of to client utils
